+++ tomcat/src/main/java/nextstep/jwp/ui/DashboardController.java
@@ -3,7 +3,7 @@ package nextstep.jwp.ui;
 import java.util.Optional;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
-import org.apache.coyote.http11.ResponseEntity;
+import org.apache.coyote.http11.response.ResponseEntity;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.request.RequestMethod;
 import org.apache.coyote.http11.response.HttpStatus;
