+++ tomcat/src/main/java/org/apache/util/UrlUtil.java
@@ -1,8 +1,5 @@
 package org.apache.util;
 
-import java.util.Arrays;
-import java.util.stream.Collectors;
-
 public class UrlUtil {
     public static String joinUrl(String... paths) {
         return String.join("/", paths)
