+++ tomcat/src/test/java/org/apache/mvc/RequestHandlerMethodTest.java
@@ -4,10 +4,11 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import org.apache.coyote.http11.ResponseEntity;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpStatus;
+import org.apache.coyote.http11.response.ResponseEntity;
 import org.apache.exception.TempException;
+import org.apache.mvc.handlerchain.RequestHandlerMethod;
 import org.junit.jupiter.api.Test;
 
 class RequestHandlerMethodTest {
